Rename packages and fix console input
