+++ src/main/java/life/qbic/io/commandline/OpenBISPasswordParser.java
@@ -1,39 +1,17 @@
 package life.qbic.io.commandline;
 
-import java.io.BufferedReader;
 import java.io.Console;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 public class OpenBISPasswordParser {
 
-  private static final Logger LOG = LogManager.getLogger(OpenBISPasswordParser.class);
-
   /**
-   * Retrieve the password from input stream
+   * Retrieve the password from the system console
    *
-   * @return the password read from the input stream
+   * @return the password read from the system console input
    */
-  public static String readPasswordFromInputStream(InputStream inputStream) {
-    String password = "";
-    try(BufferedReader inputReader = new BufferedReader(new InputStreamReader(inputStream))){
-      password = inputReader.readLine();
-    } catch (IOException e){
-      LOG.error(e);
-    }
-    return password;
-  }
-
-  /** Definition of some useful enum types for the cmd attributes */
-  public enum Attribute {
-    HELP,
-    USERNAME,
-    ID,
-    FILE,
-    BUFFER_SIZE,
-    CONSERVE_PATH
+  public static String readPasswordFromConsole() {
+    Console console = System.console();
+    char[] passwordChars = console.readPassword();
+    return String.valueOf(passwordChars);
   }
 }
