+++ src/main/java/life/qbic/model/units/Bytes.java
@@ -1,4 +1,4 @@
-package life.qbic.model.UnitConverter;
+package life.qbic.model.units;
 
 class Bytes implements UnitDisplay {
