+++ src/main/java/life/qbic/model/units/GigaBytes.java
@@ -1,4 +1,4 @@
-package life.qbic.model.UnitConverter;
+package life.qbic.model.units;
 
 class GigaBytes implements UnitDisplay {
